refactor(aviation_safety_network): Move common function to shared module
